scripts/requirements.txt
openai
psycopg2-binary
pypdf
python-dotenv